test_main.py
import pytest

from main import Board  # Replace 'your_module' with the actual module name


def test_board_initialization():
    # Test initialization without state
    board = Board()
    assert isinstance(board, Board)
    assert all(value is None for value in [cell.value for cell in board.iter_cells()])
    assert board.is_solved() is False

    # Test initialization with valid state
    valid_state = ".1..7..5.8..1..7434...3.2..7...13..6..16.....368..4.79.873.596.93..2.8....589.4.7"
    board_with_state = Board(valid_state)
    assert isinstance(board_with_state, Board)
    # Add more assertions based on valid_state

    # Test initialization with invalid state
    invalid_state = "invalid state string"
    with pytest.raises(Board.BoardException):
        Board(invalid_state)


def test_set_and_get_cell():
    board = Board()
    board.set_cell(0, 0, 5)
    assert board.get_cell_str(0, 0) == '5'
    assert board.get_cell(0, 0) == 5


def test_clear_board():
    board = Board()
    board.set_cell(0, 0, 5)
    board.clear_board()
    assert all(value is None for value in [cell.value for cell in board.iter_cells()])


def test_clone():
    board = Board(".1..7..5.8..1..7434...3.2..7...13..6..16.....368..4.79.873.596.93..2.8....589.4.7")
    cloned_board = board.clone()
    assert cloned_board.board == board.board


def test_copy_from():
    board1 = Board(".1..7..5.8..1..7434...3.2..7...13..6..16.....368..4.79.873.596.93..2.8....589.4.7")
    board2 = Board("..7434...3.79.873.596.93..2.8....589.4.7.2..7...13..6..16.....368..4.1..7..5.8..1")
    assert board1.board != board2.board
    board1.copy_from(board2)
    assert board1.board == board2.board


def test_solve():
    board = Board(".1..7..5.8..1..7434...3.2..7...13..6..16.....368..4.79.873.596.93..2.8....589.4.7")
    # Setup board in a solvable state
    board.solve_sudoku()
    assert board.is_solved()

def test_create_from_scratch():
    board = Board()
    board.solve_sudoku()
    assert board.is_solved()

def test_is_valid():
    board = Board()
    # Various cases to test
    assert board.is_valid()  # Assuming the initial board is valid
    board = Board("11..7..5.8..1..7434...3.2..7...13..6..16.....368..4.79.873.596.93..2.8....589.4.7")
    assert not board.is_valid()


def test_board_manipulations():
    board = Board()
    # Test get_row, get_column, get_chunk
    # Setup specific board state and test these methods
